Move domain images for spawner form construction
There is no longer a separate FormService, so putting the models
returned by the image service in controller.models.domain.form felt
weird. Rename that module to controller.models.domain.image since it
is used by the image service.
