+++ controller/src/controller/services/image.py
@@ -12,7 +12,7 @@ from structlog.stdlib import BoundLogger
 from ..constants import IMAGE_REFRESH_INTERVAL
 from ..exceptions import UnknownDockerImageError
 from ..models.domain.docker import DockerReference
-from ..models.domain.form import MenuImage, MenuImages
+from ..models.domain.image import MenuImage, MenuImages
 from ..models.domain.kubernetes import KubernetesNodeImage
 from ..models.domain.rspimage import RSPImage, RSPImageCollection
 from ..models.domain.rsptag import RSPImageType
