+++ controller/src/controller/services/source/base.py
@@ -8,7 +8,7 @@ from collections.abc import Mapping
 from structlog.stdlib import BoundLogger
 
 from ...models.domain.docker import DockerReference
-from ...models.domain.form import MenuImage
+from ...models.domain.image import MenuImage
 from ...models.domain.kubernetes import KubernetesNodeImage
 from ...models.domain.rspimage import RSPImage, RSPImageCollection
 from ...models.v1.prepuller import PrepulledImage
